Change one-liners to proper if-statements
The one-liner conditional check that uses &&, fails and returns 1
if the condition is not met and there is no ||.
If that conditional check happens to be on the last line in a function
and the script is run with 'set -e', this will cause the whole
function to fail which will terminate the script with exit code 1.
A proper if-then-fi statement does not fail if the condition is not met.